simplify for tutorial
